chore: tạo script đóng gói và cài đặt

+++ nova.bat
@@ -6,6 +6,15 @@ set DIR=%~dp0
 set JAVA_CMD=java
 set JAVAC_CMD=javac
 
+:: Check bundled JRE first
+if exist "%DIR%jre\bin\java.exe" (
+    set JAVA_CMD="%DIR%jre\bin\java.exe"
+    if exist "%DIR%jre\bin\javac.exe" (
+        set JAVAC_CMD="%DIR%jre\bin\javac.exe"
+    )
+    goto :check_args
+)
+
 if defined JAVA_HOME (
     set JAVA_CMD="%JAVA_HOME%\bin\java.exe"
     set JAVAC_CMD="%JAVA_HOME%\bin\javac.exe"
@@ -55,12 +64,16 @@ if "%~1"=="test-core" (
     goto :EOF
 )
 
-:: Biên dịch ngầm nếu thư mục out chưa tồn tại
-if not exist "%DIR%out" (
-    echo [Nova CLI] Khoi tao build lan dau...
-    mkdir "%DIR%out"
-    %JAVAC_CMD% -encoding UTF-8 -d "%DIR%out" -sourcepath "%DIR%src" "%DIR%src\Main.java"
+:: Chạy từ nova.jar nếu đã đóng gói, ngược lại chạy từ thư mục out
+if exist "%DIR%nova.jar" (
+    %JAVA_CMD% -cp "%DIR%nova.jar" Main %*
+) else (
+    if not exist "%DIR%out" (
+        echo [Nova CLI] Khoi tao build lan dau...
+        mkdir "%DIR%out"
+        %JAVAC_CMD% -encoding UTF-8 -d "%DIR%out" -sourcepath "%DIR%src" "%DIR%src\Main.java"
+    )
+    %JAVA_CMD% -cp "%DIR%out" Main %*
 )
 
-%JAVA_CMD% -cp "%DIR%out" Main %*
 endlocal
